Refactor location service
