+++ app/src/main/java/tice/dagger/modules/android/TICEViewModelModule.kt
@@ -35,11 +35,6 @@ abstract class TICEViewModelModule {
     @ViewModelKey(GroupMapViewModel::class)
     abstract fun bindMapViewModel(myViewModel: GroupMapViewModel): ViewModel
 
-//    @Binds
-//    @IntoMap
-//    @ViewModelKey(MapboxContainerViewModel::class)
-//    abstract fun bindMapboxContainerViewModel(myViewModel: MapboxContainerViewModel): ViewModel
-
     @Binds
     @IntoMap
     @ViewModelKey(SettingsViewModel::class)
