+++ app/src/main/java/tice/managers/LocationManagerType.kt
@@ -1,8 +1,5 @@
 package tice.managers
 
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.SharedFlow
 import tice.models.Location
 import java.lang.ref.WeakReference
