+++ app/src/playstore/AndroidManifest.xml
@@ -29,6 +29,13 @@
                 <action android:name="com.google.firebase.INSTANCE_ID_EVENT" />
             </intent-filter>
         </service>
+
+        <service
+            android:name="tice.managers.services.GMSLocationService"
+            android:enabled="true"
+            android:exported="false"
+            android:foregroundServiceType="location"
+            android:stopWithTask="true" />
     </application>
 
 </manifest>
